add : Code form && rename model to DTO
