+++ requirements.txt
@@ -1,3 +1,4 @@
 Flask==1.1.2
 asyncio==3.4.3
 jwt==1.2.0
+SQLAlchemy==1.3.20
