+++ src/misc/__init__.py
@@ -1,3 +1,4 @@
 from .config import *
 from .util import *
-from .model import *
+from .DTO import *
+from .db import *
